+++ src/com/intellij/idea/plugin/hybris/flexibleSearch/settings/FlexibleSearchSettings.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
+ * Copyright (C) 2023 EPAM Systems <hybrisideaplugin@epam.com>
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -30,6 +30,7 @@ data class FlexibleSearchSettings(
 
     var completion: FlexibleSearchCompletionSettings = FlexibleSearchCompletionSettings(),
     var folding: FlexibleSearchFoldingSettings = FlexibleSearchFoldingSettings(),
+    var documentation: FlexibleSearchDocumentationSettings = FlexibleSearchDocumentationSettings(),
 )
 
 data class FlexibleSearchCompletionSettings(
@@ -38,6 +39,11 @@ data class FlexibleSearchCompletionSettings(
     var suggestTableAliasNames: Boolean = true,
     var injectCommaAfterExpression: Boolean = true,
     var defaultTableAliasSeparator: String = HybrisConstants.FXS_TABLE_ALIAS_SEPARATOR_DOT,
+)
+
+data class FlexibleSearchDocumentationSettings(
+    var enabled: Boolean = true,
+    var showTypeDocumentation: Boolean = true,
 )
 
 class FlexibleSearchFoldingSettings : BaseState() {