+++ src/com/intellij/idea/plugin/hybris/flexibleSearch/settings/FlexibleSearchSettingsConfigurableProvider.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
+ * Copyright (C) 2023 EPAM Systems <hybrisideaplugin@epam.com>
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -49,6 +49,7 @@ class FlexibleSearchSettingsConfigurableProvider(val project: Project) : Configu
         private val reservedWordsModel = EnumComboBoxModel(ReservedWordsCase::class.java)
         private val tableAliasSeparatorsModel = CollectionComboBoxModel(listOf(".", ":"))
         private lateinit var foldingEnableCheckBox: JCheckBox
+        private lateinit var documentationEnableCheckBox: JCheckBox
 
         override fun apply() {
             super.apply()
@@ -136,6 +137,23 @@ class FlexibleSearchSettingsConfigurableProvider(val project: Project) : Configu
                         .bindSelected(state.folding::showLanguageForYColumn)
                         .enabledIf(foldingEnableCheckBox.selected)
                         .comment("If checked localized attribute <code>{name[en]}</code> will be represented as <code>name:en</code>")
+                }
+            }
+            group("Documentation") {
+                row {
+                    documentationEnableCheckBox = checkBox("Enable code folding")
+                        .bindSelected(state.documentation::enabled)
+                        .component
+                }
+                row {
+                    checkBox("Show documentation for type")
+                        .comment(
+                            """
+                            When enabled short description of the type will be shown on-hover as a tooltip for type used in the <code>FROM {}</code> clause.
+                        """.trimIndent()
+                        )
+                        .bindSelected(state.documentation::showTypeDocumentation)
+                        .enabledIf(documentationEnableCheckBox.selected)
                 }
             }
         }